requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg==3.1.12
python-dotenv==1.0.0
Werkzeug==2.3.7